+++ invert/solvers/beamformer.py
@@ -17,9 +17,9 @@ class SolverMVAB(BaseSolver):
     beamforming design. Signal Processing, 93(12), 3264-3277.
 
     '''
-    def __init__(self, name="Minimum Variance Adaptive Beamformer", **kwargs):
+    def __init__(self, name="Minimum Variance Adaptive Beamformer", reduce_rank=True, rank="auto", **kwargs):
         self.name = name
-        return super().__init__(**kwargs)
+        return super().__init__(reduce_rank=reduce_rank, rank=rank, **kwargs)
 
     def make_inverse_operator(self, forward, mne_obj, *args, alpha='auto', **kwargs):
         ''' Calculate inverse operator.
@@ -75,9 +75,9 @@ class SolverLCMV(BaseSolver):
     approach to spatial filtering. IEEE assp magazine, 5(2), 4-24.
     
     '''
-    def __init__(self, name="LCMV Beamformer", **kwargs):
+    def __init__(self, name="LCMV Beamformer", reduce_rank=True, rank="auto", **kwargs):
         self.name = name
-        return super().__init__(**kwargs)
+        return super().__init__(reduce_rank=reduce_rank, rank=rank, **kwargs)
 
     def make_inverse_operator(self, forward, mne_obj, *args, alpha='auto', weight_norm=True, verbose=0, **kwargs):
         ''' Calculate inverse operator.
@@ -110,15 +110,16 @@ class SolverLCMV(BaseSolver):
         y -= y.mean(axis=0)
         
         I = np.identity(n_chans)
-
+        C = y@y.T
+        
         # Recompute regularization based on the max eigenvalue of the Covariance
         # Matrix (opposed to that of the leadfield)
         self.alphas = np.logspace(-4, 1, self.n_reg_params) * np.diagonal(y@y.T).mean()
 
         inverse_operators = []
         for alpha in self.alphas:
-            C = y@y.T + alpha*I
-            C_inv = np.linalg.inv(C)
+            
+            C_inv = np.linalg.inv(C + alpha*I)
 
             W = (C_inv @ leadfield) / np.diagonal(leadfield.T @ C_inv @ leadfield)
 
@@ -129,6 +130,11 @@ class SolverLCMV(BaseSolver):
 
         self.inverse_operators = [InverseOperator(inverse_operator, self.name) for inverse_operator in inverse_operators]
         return self
+
+
+
+    
+        
 
 class SolverSMV(BaseSolver):
     ''' Class for the Standardized Minimum Variance (SMV) Beamformer inverse
@@ -144,9 +150,9 @@ class SolverSMV(BaseSolver):
     Biomedical Signal Processing and Control, 14, 175-188.
 
     '''
-    def __init__(self, name="SMV Beamformer", **kwargs):
+    def __init__(self, name="SMV Beamformer", reduce_rank=True, rank="auto", **kwargs):
         self.name = name
-        return super().__init__(**kwargs)
+        return super().__init__(reduce_rank=reduce_rank, rank=rank, **kwargs)
 
     def make_inverse_operator(self, forward, mne_obj, *args, weight_norm=True, alpha='auto', **kwargs):
         ''' Calculate inverse operator.
@@ -215,9 +221,9 @@ class SolverWNMV(BaseSolver):
     Biomedical Signal Processing and Control, 14, 175-188.
 
     '''
-    def __init__(self, name="WNMV Beamformer", **kwargs):
+    def __init__(self, name="WNMV Beamformer", reduce_rank=True, rank="auto", **kwargs):
         self.name = name
-        return super().__init__(**kwargs)
+        return super().__init__(reduce_rank=reduce_rank, rank=rank, **kwargs)
 
     def make_inverse_operator(self, forward, mne_obj, *args, weight_norm=True, alpha='auto', **kwargs):
         ''' Calculate inverse operator.
@@ -286,9 +292,9 @@ class SolverHOCMV(BaseSolver):
     Biomedical Signal Processing and Control, 14, 175-188.
 
     '''
-    def __init__(self, name="HOCMV Beamformer", **kwargs):
+    def __init__(self, name="HOCMV Beamformer", reduce_rank=True, rank="auto", **kwargs):
         self.name = name
-        return super().__init__(**kwargs)
+        return super().__init__(reduce_rank=reduce_rank, rank=rank, **kwargs)
 
     def make_inverse_operator(self, forward, mne_obj, *args, weight_norm=True, alpha='auto', order=3, verbose=0, **kwargs):
         ''' Calculate inverse operator.
@@ -363,9 +369,9 @@ class SolverESMV(BaseSolver):
     Biomedical Signal Processing and Control, 14, 175-188.
     
     '''
-    def __init__(self, name="ESMV Beamformer", **kwargs):
+    def __init__(self, name="ESMV Beamformer", reduce_rank=True, rank="auto", **kwargs):
         self.name = name
-        return super().__init__(**kwargs)
+        return super().__init__(reduce_rank=reduce_rank, rank=rank, **kwargs)
 
     def make_inverse_operator(self, forward, mne_obj, *args, alpha='auto', **kwargs):
         ''' Calculate inverse operator.
@@ -402,28 +408,47 @@ class SolverESMV(BaseSolver):
         # Matrix (opposed to that of the leadfield)
         C = y@y.T
         self.alphas = self.get_alphas(reference=C)
+        subspace = self.reduce_rank_matrix(C)
 
-        U, s, _ = np.linalg.svd(C)
-        # j = find_corner(np.arange(len(s)), s)
         
-        # Find number of Signal Subspaces:
-        j = np.where(((s**2)*len((s**2)) / (s**2).sum()) < np.exp(-16))[0][0]
-
-        Us = U[:, :j]
-        # Un = U[:, j:]
-
         inverse_operators = []
         for alpha in self.alphas:
             C_inv = np.linalg.inv(C + alpha * I)
 
             W_mv = (C_inv @ leadfield) / np.diagonal(leadfield.T @ C_inv @ leadfield)
-            W = Us @ Us.T @ W_mv
+            W = subspace @ W_mv
 
             inverse_operator = W.T
             inverse_operators.append(inverse_operator)
 
         self.inverse_operators = [InverseOperator(inverse_operator, self.name) for inverse_operator in inverse_operators]
         return self
+
+    def reduce_rank_matrix(self, C):
+        # Calculte eigenvectors and eigenvalues
+        U, s, _ = np.linalg.svd(C, full_matrices=False)
+        
+        # Find optimal rank using L-curve Corner approach:
+        n_comp_l = find_corner(np.arange(len(s)), s)
+        
+        # Find optimal rank using drop-off approach:
+        s_ = s / s.max()
+        n_comp_d = np.where( abs(np.diff(s_)) < 0.001 )[0]
+        if len(n_comp_d) > 0:
+            n_comp_d = n_comp_d[0] + 2
+        else:
+            n_comp_d = n_comp_l
+        
+        # Kaiser Rule
+        n_comp_k = np.where(s < np.mean(s))[0][0]
+        print(n_comp_k, n_comp_l, n_comp_d)
+        
+        # Combine the approaches
+        n_comp = np.ceil((n_comp_d + n_comp_l + n_comp_k)/3).astype(int)
+        
+        # Transform data
+        subspace = U[:, :n_comp] @ U[:, :n_comp].T
+        return subspace
 
 class SolverMCMV(BaseSolver):
     ''' Class for the Multiple Constrained Minimum Variance (MCMV) Beamformer
@@ -440,9 +465,9 @@ class SolverMCMV(BaseSolver):
     (MCMV) performance in connectivity analyses. NeuroImage, 208, 116386.
 
     '''
-    def __init__(self, name="MCMV Beamformer", **kwargs):
+    def __init__(self, name="MCMV Beamformer", reduce_rank=True, rank="auto", **kwargs):
         self.name = name
-        return super().__init__(**kwargs)
+        return super().__init__(reduce_rank=reduce_rank, rank=rank, **kwargs)
 
     def make_inverse_operator(self, forward, mne_obj, *args, weight_norm=True, noise_cov=None, alpha='auto', verbose=0, **kwargs):
         ''' Calculate inverse operator.
@@ -489,11 +514,92 @@ class SolverMCMV(BaseSolver):
         for alpha in self.alphas:
             C_inv = np.linalg.inv(C + alpha * I)
 
-            # W = C_inv @ leadfield * np.diagonal(np.linalg.inv(leadfield.T @ C_inv @ leadfield))
-            W = np.dot(leadfield.T, np.dot(C_inv, leadfield))
-            W = np.dot(np.linalg.inv(W + leadfield.T @ noise_cov @ leadfield), np.dot(leadfield.T, C_inv)).T
+            W = C_inv @ leadfield * np.diagonal(np.linalg.inv(leadfield.T @ C_inv @ leadfield))
+            # W = C_inv @ leadfield @ np.linalg.inv(leadfield.T @ C_inv @ leadfield)
+            
+            # W = np.dot(leadfield.T, np.dot(C_inv, leadfield))
+            # W = np.dot(np.linalg.inv(W + leadfield.T @ noise_cov @ leadfield), np.dot(leadfield.T, C_inv)).T
     
 
+            if self.weight_norm:
+                W /= np.linalg.norm(W, axis=0)
+
+            inverse_operator = W.T
+            inverse_operators.append(inverse_operator)
+
+        self.inverse_operators = [InverseOperator(inverse_operator, self.name) for inverse_operator in inverse_operators]
+        return self
+
+    
+
+class SolverHOCMCMV(BaseSolver):
+    ''' Class for the Higher-Order Covariance Multiple Constrained Minimum Variance (HOCMCMV)
+        Beamformer inverse solution [1].
+    
+    Attributes
+    ----------
+    forward : mne.Forward
+        The mne-python Forward model instance.
+    
+    References
+    ----------
+    [1] Jonmohamadi, Y., Poudel, G., Innes, C., Weiss, D., Krueger, R., & Jones,
+    R. (2014). Comparison of beamformers for EEG source signal reconstruction.
+    Biomedical Signal Processing and Control, 14, 175-188.
+
+    '''
+    def __init__(self, name="HOCMCMV Beamformer", reduce_rank=True, rank="auto", **kwargs):
+        self.name = name
+        return super().__init__(reduce_rank=reduce_rank, rank=rank, **kwargs)
+
+    def make_inverse_operator(self, forward, mne_obj, *args, weight_norm=True, alpha='auto', order=3, verbose=0, **kwargs):
+        ''' Calculate inverse operator.
+
+        Parameters
+        ----------
+        forward : mne.Forward
+            The mne-python Forward model instance.
+        mne_obj : [mne.Evoked, mne.Epochs, mne.io.Raw]
+            The MNE data object.
+        weight_norm : bool
+            Normalize the filter weight matrix W to unit length of the columns.
+        alpha : float
+            The regularization parameter.
+        order : int
+            The order of the covariance matrix. Should be a positive integer not
+            evenly divisible by two {3, 5, 7, ...}
+        
+        Return
+        ------
+        self : object returns itself for convenience
+        '''
+        super().make_inverse_operator(forward, *args, alpha=alpha, **kwargs)
+        data = self.unpack_data_obj(mne_obj)
+
+        leadfield = self.leadfield
+        leadfield -= leadfield.mean(axis=0)
+        leadfield /= np.linalg.norm(leadfield, axis=0)
+        n_chans, n_dipoles = self.leadfield.shape
+        
+        self.weight_norm = weight_norm
+        
+        y = data
+        y -= y.mean(axis=0)
+        I = np.identity(n_chans)
+
+        # Recompute regularization based on the max eigenvalue of the Covariance
+        # Matrix (opposed to that of the leadfield)
+        C = y@y.T
+        self.alphas = self.get_alphas(reference=C)
+        
+        inverse_operators = []
+        for alpha in self.alphas:
+            C_inv = np.linalg.inv(C + alpha * I)
+            C_inv_n = deepcopy(C_inv)
+            for _ in range(order-1):
+                C_inv_n = np.linalg.inv(C_inv_n)
+            W = C_inv @ leadfield * np.diagonal(np.linalg.inv(leadfield.T @ C_inv_n @ leadfield))
+            
             if self.weight_norm:
                 W /= np.linalg.norm(W, axis=0)
 
@@ -518,9 +624,9 @@ class SolverReciPSIICOS(BaseSolver):
     with correlated sources. Neuroimage, 228, 117677.
 
     '''
-    def __init__(self, name="ReciPSIICOS", **kwargs):
+    def __init__(self, name="ReciPSIICOS", reduce_rank=True, rank="auto", **kwargs):
         self.name = name
-        return super().__init__(**kwargs)
+        return super().__init__(reduce_rank=reduce_rank, rank=rank, **kwargs)
 
     def make_inverse_operator(self, forward, mne_obj, *args, weight_norm=True, alpha='auto', verbose=0, **kwargs):
         ''' Calculate inverse operator.
@@ -609,9 +715,9 @@ class SolverSAM(BaseSolver):
     aperture magnetometry (SAM). Recent advances in biomagnetism.
 
     '''
-    def __init__(self, name="SAM Beamformer", **kwargs):
+    def __init__(self, name="SAM Beamformer", reduce_rank=True, rank="auto", **kwargs):
         self.name = name
-        return super().__init__(**kwargs)
+        return super().__init__(reduce_rank=reduce_rank, rank=rank, **kwargs)
 
     def make_inverse_operator(self, forward, mne_obj, *args, weight_norm=True, alpha='auto', verbose=0, **kwargs):
         ''' Calculate inverse operator.